git-svn-id: http://192.168.0.243:4435/svn/bee/branches/alterna@2266 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/server/modules/ec/EcModuleBean.java
@@ -2403,8 +2403,8 @@ public class EcModuleBean implements BeeModule {
         if (!itemSuppliers.isEmpty()) {
           item.setSuppliers(itemSuppliers);
 
-          int primaryStock = 0;
-          int secondaryStock = 0;
+          double primaryStock = 0;
+          double secondaryStock = 0;
 
           for (ArticleSupplier itemSupplier : itemSuppliers) {
             if (!primaryWarehouses.isEmpty()) {
